tweaked the logic about a bit
Tweaked it to follow drop-loot-tables's KO dropper somewhat:
1. Roll for form's loot from mod logic.
2. If there isn't a form-specific loot table from mod:
  a. Roll for form's loot from Cobblemon.
  b. Add Cobblemon's drops to the list.
It's a bit of an inverse, to save on not needing to roll if there is a form-specific loot table from the mod, whereas KO dropper had to deal with already having the drops rolled, and agreed, we are cajolling the results into an ItemStack list rather than the other way around.

Also kept it so that we *only* add items from Cobblemon's drops (according to the ItemDropEntry's internal logic), *but also* allow other types of drops to still occur, if possible, given they do have a position and a level, but not a player or Pokemon entity.

Keeping it out of PastureCollectorDropper as a Dropper's scope is just that layer over grabbing arbitrary loot tables and rolling on them.